n64: fix CPU core to correctly handle 64-bit PC
PC register in VR4300 is 64-bit but there were many places where
the register was incorrectly treated as a 32-bit one, effectively
discarding the higher half.

Fix several tests in n64-systemtest ROM.

+++ ares/n64/cpu/cpu.hpp
@@ -41,7 +41,7 @@ struct CPU : Thread {
   auto power(bool reset) -> void;
 
   struct Pipeline {
-    u32 address;
+    u64 address;
     u32 instruction;
 
     struct InstructionCache {
@@ -65,7 +65,7 @@ struct CPU : Thread {
 
     auto inDelaySlot() const -> bool { return state == DelaySlot; }
     auto reset() -> void { state = Step; }
-    auto take(u32 address) -> void { state = Take; pc = address; }
+    auto take(u64 address) -> void { state = Take; pc = address; }
     auto delaySlot() -> void { state = DelaySlot; }
     auto exception() -> void { state = Exception; }
     auto discard() -> void { state = Discard; }
